refactor profile link header
