+++ src/app/[link]/page.tsx
@@ -9,7 +9,7 @@ import { api } from "@/trpc/server";
 import BentoCard from "@/components/bento/card";
 import BentoLayout from "@/components/bento/layout";
 import ActionBar from "@/components/bento/action-bar";
-import ProfileLinkEditor from "@/components/profile-link-editor";
+import ProfileLinkHeader from "./_components/header";
 
 type Props = {
   params: {
@@ -57,12 +57,12 @@ export default async function Page({ params }: Props) {
     );
   }
 
-  await api.profileLink.recordVisit.mutate({ id: profileLink.id });
+  void api.profileLink.recordVisit.mutate({ id: profileLink.id });
 
   return (
     <div className="h-full w-full max-w-3xl">
       <div className="flex flex-col gap-y-6">
-        <ProfileLinkEditor profileLink={profileLink} />
+        <ProfileLinkHeader profileLink={profileLink} />
 
         <BentoLayout profileLink={profileLink}>
           {profileLink.bento.map((b) => (
