+++ src/trpc/react.ts
@@ -1,5 +1,10 @@
 import { createTRPCReact } from "@trpc/react-query";
+import { type inferRouterInputs, type inferRouterOutputs } from "@trpc/server";
 
 import { type AppRouter } from "@/server/api/root";
 
 export const api = createTRPCReact<AppRouter>({});
+
+export type RouterInputs = inferRouterInputs<AppRouter>;
+
+export type RouterOutputs = inferRouterOutputs<AppRouter>;
